scala: Check file exists on session create
The native create session call returns null in that case, using it results in native segfault.

Guard against that specific file not found failure by checking for file exists, also add checks on the impl classes to require non-null pointers to fail fast in the case where the calling logic allows a null to get through.

Removes logic from the gRPC example that pre allocates empty space in the session.  That logic was added as a result of misunderstanding the api bug.