out-of-band file change handling (#637)
This is done by creating an initial "session checkpoint" which
copies the original file to a configured checkpoint directory
and uses that for edit modeling for the lifetime of the session.
Once the session is destroyed, it will cleanup all checkpoint
files including this initial session checkpoint file.  This way
changes made to the orginal file will not interfere with an
editing session opened on the same file.